[bug] Fix scrolling issue with textarea
We decided to fix this by resizing the textarea based on its content.
This part was simple, but there was a weird quirk with how the textarea
would love to "snap" to its position when typing at the bottom of the
input while the element had just resized.

We worked around this part by completely removing the bottom margin to
the textarea.

Closes #8